// src/domain/models/service/Service.ts
export interface Service {
  id: string
  name: string
  description: string
  price: number
  duration_min: number
  img: string
  employees: Employee[]
  days: Day[]
  hourStart: string
  hourEnd: string
}

export interface Employee {
  id: number,
  name: string,
  lastName: string,
  img: string,
  role: Role,
  status: EmployeeStatus
}

export type Day =
  | "lunes"
  | "martes"
  | "miercoles"
  | "jueves"
  | "viernes"
  | "sabado"
  | "domingo";

export const employeeStatuses = {
  ACTIVE: "active",
  VACATION: "vacation",
  SICK: "sick",
  DAY_OFF: "day off"

} as const
export type EmployeeStatus = typeof employeeStatuses[keyof typeof employeeStatuses]

export const roles = {
  ADMIN: "admin",
  EMPLOYEE: "employee",
  USER: "user"
} as const

export type Role = typeof roles[keyof typeof roles]